LibCompress has been trashed! The GPL is an UNACCEPTABLE license of anything license related under all circumstances.
The GPL is a viral license, and in this particular case incompatible with "zlib as long as no story script has been touched" license system and that is precisely what this can be used for, once this is fully being put to the work this is going to be used for! A game entirely into zlib is incompatible with GPL (a GPL source calling a zlib source is a different matter), and there's a reason my story related stuff has a pretty restrictive license.

I did find a good replacement, however this DOES require a c-source, but I suppose with luarocks, this can be helped... somehow... These kinds of things are always pretty sensitive, and I need to see how this will work out in Windows. Windows is a pretty headstrong system after all!