Read dir
It should be noted that I just manually translated the routines I wrote in Go.
Path support is not there for REQUIRE and IMPORT yet (issued as #1), but that is no short term thing to be done (I never used it myself, actually. I just built it in to be sure support was there if needed or in demand)

Nothing has been tested yet, so I'm prepared for a big mess!